test: rename & split internal module tests (#37318)
